+++ src/registers/doorbell.rs
@@ -1,7 +1,7 @@
 //! Doorbell Register
 
 use bit_field::BitField;
-use core::convert::TryInto;
+use core::{convert::TryInto, fmt};
 
 /// The element of the Doorbell Array.
 #[repr(transparent)]
@@ -16,5 +16,12 @@ impl Register {
     /// Set a doorbell target.
     pub fn set_doorbell_target(&mut self, target: u8) {
         self.0.set_bits(0..=7, target.into());
+    }
+}
+impl fmt::Debug for Register {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_struct("doorbell::Register")
+            .field("doorbell_target", &self.doorbell_target())
+            .finish()
     }
 }